routes according to doc for review and removed dummy text
