--- src/pages/reviews/[id]/index.tsx
@@ -1,36 +1,0 @@
-import Head from "next/head";
-import SendReview from "@/components/SendReview";
-
-export default function MovieDetailsPage({}: {}) {
-  return (
-    <>
-      <Head>
-        <title>Lule Northern Light Cinema</title>
-        <meta
-          name="This page get you see what screenings this movie has and information about it"
-          content="Kino project in next.js"
-        />
-        <meta name="viewport" content="width=device-width, initial-scale=1" />
-      </Head>
-      <SendReview />
-      <h1>hanne</h1>
-      Lorem ipsum dolor sit amet consectetur adipisicing elit. Incidunt a quia
-      consequatur quasi nisi quis vero dolorem totam ullam quisquam laborum
-      similique eveniet, eius temporibus voluptatem quaerat? Quas, at
-      accusantium! Modi perferendis enim ratione aut eum officiis alias
-      reprehenderit rem, voluptatibus repellendus doloremque nisi totam expedita
-      delectus dolore quisquam hic excepturi. Consectetur eum magnam autem sit
-      ratione. Nesciunt, ipsa illum. Cum sint unde voluptatibus eos ducimus in,
-      accusantium id nobis officiis minima debitis laborum iste. Iste
-      consequuntur deleniti repellendus qui mollitia aliquam perspiciatis
-      molestiae illo iusto, dolorum quisquam atque. Ipsa. Illo fugit omnis
-      doloribus laboriosam ipsum debitis amet dolore perspiciatis officia
-      repudiandae iusto alias sint voluptas, nobis unde et fuga, dicta in?
-      Beatae voluptatem, nobis praesentium quia nulla iure adipisci. Expedita a
-      natus velit consequatur tenetur maiores, delectus cupiditate debitis
-      similique dolorum magni reiciendis incidunt ea vero dolore aut iure
-      recusandae voluptas, iste excepturi, veritatis vitae architecto deleniti?
-      Animi, obcaecati.
-    </>
-  );
-}
